+++ server/src/version.rs
@@ -1,0 +1,1 @@
+pub const VERSION: &'static str = "v0.4.0-6-gd767e94";
